[ObjC] Follow-up r350768 and allow the use of unavailable methods that are declared in a parent class from within the @implementation context
This commit extends r350768 and allows the use of methods marked as unavailable
that are declared in a parent class/category from within the @implementation of
the class where the method is marked as unavailable.
This allows users to call init that's marked as unavailable even if they don't
define it.

rdar://47134898

Differential Revision: https://reviews.llvm.org/D56816


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@351459 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit f812df94335cccf585168fcda36d6b683b625640)

apple-llvm-split-commit: 01e6db1629e0a5540ec982006e73351ee463f656
apple-llvm-split-dir: clang/