[AArch64] Add 'apple-latest' CPU alias
The 'apple-latest' alias is supposed to provide a CPU that contains the
latest Apple processor model supported by LLVM.

This is supposed to be used by tools like lldb to provide a target that
supports most of the CPU features.

For now, this is mapped to Cyclone.

Differential Revision: https://reviews.llvm.org/D56384

rdar://45147519

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@352412 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: def803f04bbcf21126e1619aa00057d394008faa
apple-llvm-split-dir: llvm/