[analyzer] Canonicalize declarations within variable regions.
Memory region that correspond to a variable is identified by the variable's
declaration and, in case of local variables, the stack frame it belongs to.

The declaration needs to be canonical, otherwise we'd have two different
memory regions that correspond to the same variable.

Fix such bug for global variables with forward declarations and assert
that no other problems of this kind happen.

Differential Revision: https://reviews.llvm.org/D57619


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@353353 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 4ec0edc0674455eab03724a948d2059f8870798a)

apple-llvm-split-commit: f555bbb0f8e4e31ec4db46983dd8e788ad216eaf
apple-llvm-split-dir: clang/