-gmodules: Don't emit incomplete breadcrumbs pointing to nonexistant PCM files.
When a module name is specified as -fmodule-name, that module gets a
clang::Module object, but it won't actually be built or imported; it
will be textual. CGDebugInfo wouldn't detect this and them emit a
DICompileUnit that had a hash but no name and that confused both
dsymutil, LLDB, and myself.

rdar://problem/47926508

Differential Revision: https://reviews.llvm.org/D57976

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@353578 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 683a8b8183a89168d845870800e30c8ee733fb12)

apple-llvm-split-commit: d7c377f94da8145f19f472ede6e0924c2266643d
apple-llvm-split-dir: clang/