[attributes] Add an attribute for server routines in Mach kernel and extensions.
The new __attribute__ ((mig_server_routine)) is going to be used for annotating
Mach Interface Generator (MIG) callback functions as such, so that additional
static analysis could be applied to their implementations. It can also be
applied to regular functions behavior of which is supposed to be identical to
that of a MIG server routine.

Differential Revision: https://reviews.llvm.org/D58365


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@354530 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 4c12fc335614018800bef94556f09aa0a50e51c6)

apple-llvm-split-commit: 12b4dab637a346f2f1f444c45a9493f975564a43
apple-llvm-split-dir: clang/