[Utility] Fix ArchSpec.MergeFrom to correctly merge environments
Summary:
This behavior was originally added in rL252264 (git commit 76a7f365da)
in order to be extra careful with handling platforms like watchos and tvos.
However, as far as triples go, those two (and others) are treated as OSes and
not environments, so that should not really apply here.

Additionally, this behavior is incorrect and can lead to incorrect ArchSpecs.
Because android is specified as an environment and not an OS, not propogating
the environment can lead to modules and targets being misidentified.

Differential Revision: https://reviews.llvm.org/D58664

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@354938 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 3e54bb6a80dcb0da215488b82c6c256ef1f9f427
apple-llvm-split-dir: lldb/