[ASTImporter] Improve import of FileID.
Summary:
Even if the content cache has a directory and filename, it may be a virtual file.
The old code returned with error in this case, but it is worth to try to handle
the file as it were a memory buffer.

Reviewers: a.sidorin, shafik, martong, a_sidorin

Reviewed By: shafik

Subscribers: efriedma, rnkovacs, cfe-commits, dkrupp, martong, Szelethus, gamesh411

Tags: #clang

Differential Revision: https://reviews.llvm.org/D57590

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@355000 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 128411fb1722b87a4bea45913b42dc511d9c1b5d)

apple-llvm-split-commit: 356c0fc7674680a4bde112facd6d27c01a889902
apple-llvm-split-dir: clang/