Fix invalid use of StringRef::data in Socket::DecodeHostAndPort
the input StringRef is not guaranteed to be null-terminated, so using
data to get the c string is wrong. Luckily, in two of the usages the
target function already accepts a StringRef so we can just drop the
data() call, and the third one is easily replaced by a stringref-aware
function.

Issue found by msan.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@355817 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit dae6031f2a1212ddd04988867aed4740a028568e)

apple-llvm-split-commit: 0d90b91f856af90ed99e9c277d94109c81248179
apple-llvm-split-dir: lldb/