[Python] Fix TestDataFormatterSmartArray to work across python versions.
Python 3 default encoding is utf-8, so taking random bytes and
interpreting them as a string might result in invalid unicode sequences.
As the only thing we care about here is that the formatter shows the
elements of the underyling array, relax the string matching (this is
good enough as all the elements are distinct so they resolve to different
strings).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@356096 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 84fc92d2a352b5b7b320e91ece8897a50baeddea
apple-llvm-split-dir: lldb/