[Reproducers] Fix GDB remote flakiness during replay
This fixes the flakiness of the GDB remote reproducer during replay. It
was caused by a combination sending one ACK to many from the replay
server and the code that "flushes" any queued GDB remote packets in
GDBRemoteCommunicationClient::HandshakeWithServer.

The spurious ACK was the result of combining both implicit and explicit
handling of ACKs in the replay server. The handshake consists of an ACK
followed by an QStartNoAckMode. As long as we haven't seen any
QStartNoAckMode, we were sending implicit acknowledgments. So the first
ACK got acknowledged twice, once implicitly, and once as part of the
replay.

The reason we didn't notice this was the code in HandshakeWithServer
that "waits for any responses that might have been queued up in the
remote GDB server and flush them all". A 10ms timeout is used to move on
when no packets are left. If the second ACK didn't make it within those
10ms, all packets were offset by one.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@356825 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 2f979654e7e62f30be20475bc33cd9a5637d80cd)

apple-llvm-split-commit: a43d73dfc6576150bd90f55e63fedcb896aea044
apple-llvm-split-dir: lldb/