[NativePDB] Don't fail on import modules.
A recent patch to LLD started emitting information about import modules.
These are represented as compile units in the PDB, but with no
additional debug info.  This was confusing the native pdb reader, who
expected that the debug info stream be present.

This should fix failing tests on the Windows bots.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@357513 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 7fa0d66a3eb2c38904c1dbfbbf52f0dc5b332a6f
apple-llvm-split-dir: lldb/