[ScriptInterpreter] Pass the debugger instead of the command interpreter
As discussed in D61090, there's no good reason for the script
interpreter to depend on the command interpreter. When looking at the
code, it becomes clear that we mostly use the command interpreter as a
way to access the debugger. Hence, it makes more sense to just pass that
to the script interpreter directly.

This is part 1 out of 2. I have another patch in the pipeline that
changes the ownership of the script interpreter to the debugger as well,
but I didn't get around to finish that today.

Differential revision: https://reviews.llvm.org/D61172

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359330 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 6ee3ba8e587eac7f698583f7da45f523f11e4238)

apple-llvm-split-commit: 8abff6b561242966c90bc83be953295aad353906
apple-llvm-split-dir: lldb/