[Modules] Make header inclusion order from umbrella dirs deterministic
Sort the headers by name before adding the includes in
collectModuleHeaderIncludes. This makes the include order for building
umbrellas deterministic across different filesystems and also guarantees
that the ASTWriter always dump top headers in the same order.

There's currently no good way to test for this behavior.

This was first introduced in r289478 and later reverted. Attempting it
again.

rdar://problem/28116411

apple-llvm-split-commit: 0f85b105e4f79ee320400c41f30416576a7fcf91
apple-llvm-split-dir: clang/