[driver][macOS] Link libarclite from the default toolchain when clang is running in a toolchain outside of xcode
'libarclite' usually lives in the same toolchain as 'clang'. However, the
Swift open source toolchains for macOS distribute Clang without 'libarclite'.
In that case, to allow the linker to find 'libarclite', we point to the
'libarclite' that should be in the XcodeDefault toolchain instead. The
path to the toolchain is inferred from the SDK path if it's specified.

https://bugs.swift.org/browse/SR-9972
rdar://49947573

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@359353 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 6c1a7472fe28f11e86b8e44e41da52660b7a5719)

apple-llvm-split-commit: c609f16f149a94687d7dc41c64bef05c2f6f5077
apple-llvm-split-dir: clang/