[SwiftASTContext] Library search paths should be consistent with the ones in repl_swift.
The general idea behind the REPL is that of creating a dummy
target for lldb to attach to run swift expressions. The dummy target
dlopen()s libswiftCore, so that people don't have to type `import Swift`
when they open it. In order to find libswiftCore, the build system sets
a couple of @rpath, giving priority to the just built stdlib (if any)
and falling back to the OS standard library (in /usr/lib/swift).

Once the REPL has been launched, if somebody tries to import, e.g.
Foundation, lldb calls SwiftASTContext::LoadLibraryUsingPaths, which
asks GetLibrarySearchPaths() where to look. GetLibrarySearchPaths()
sets the first rpath to /usr/lib/swift and then falls back to the
stdlib just built.

In this tragedy of commons, lldb ends up loading two different
libSwiftCore which might be not compatible with each other. This
results in weird crashes, because in presence of symbols implemented
in multiple libraries, the one that's picked is undefined.

<rdar://problem/50022178>

apple-llvm-split-commit: fe4e4ded87b11b4777503c4b00354de5e8e7b97b
apple-llvm-split-dir: lldb/