[CodeGen][ObjC] Emit invoke instead of call to call `objc_release` when necessary.
Prior to r349952, clang used to call objc_msgSend when sending a release
messages, emitting an invoke instruction instead of a call instruction
when it was necessary to catch an exception. That changed in r349952
because runtime function objc_release is called as a nounwind function,
which broke programs that were overriding the dealloc method and
throwing an exception from it. This patch restores the behavior prior to
r349952.

rdar://problem/50253394

Differential Revision: https://reviews.llvm.org/D61803

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@360474 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit b30d40eb8d46a8ffe901cd75582b74b92cb7b377)

apple-llvm-split-commit: 48809e95696dff880729fae60bbeae5e7acd4248
apple-llvm-split-dir: clang/